Add
GUI using tkinter for color selection and display
optimization results

In this new version of the code, we have integrated the tkinter
library to create a graphical user interface (GUI) for the color
mixer application. Users can now select the desired color using
a color picker and choose the available paint colors from a list.
The optimized color mix proportions are displayed on the
interface along with the final mixed color.